bullet code simplification
