--- Assets/Prefabs/Monster Bullet.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5b1f6053672c122488c1cdd1805aefc7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
