--- Assets/Scripts/GameScripts/CollisionDamage.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f6795160c9fc31e4684b7704d6f4d1b8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
